Added bowtie2 module
